refactor: change structure of the project fix ESLint dependencies
