+++ src/components/ErrorBoundary.tsx
@@ -1,4 +1,3 @@
-// src/components/ErrorBoundary.tsx
 import { Component, ErrorInfo, ReactNode } from 'react';
 
 interface Props {
