+++ src/components/Results.tsx
@@ -1,26 +1,17 @@
-// src/components/Results.tsx
 import { Component } from 'react';
-
-interface Item {
-  uid: string;
-  name: string;
-}
-
-interface Props {
-  items: Item[];
-}
+import { Props } from '../types/SearchResult.ts';
 
 class Results extends Component<Props> {
   render() {
     const { items } = this.props;
     return (
-      <div>
+      <section>
         {items.map((item) => (
           <div key={item.uid}>
             <h3>{item.name}</h3>
           </div>
         ))}
-      </div>
+      </section>
     );
   }
 }
