+++ src/types/App.ts
@@ -1,0 +1,7 @@
+import { SearchItem } from './SearchResult.ts';
+
+export interface State {
+  items: SearchItem[];
+  loading: boolean;
+  error: Error | null;
+}
