+++ src/types/SearchBar.ts
@@ -1,0 +1,9 @@
+import { State as AppState } from './App.ts';
+
+export interface Props {
+  onSearch: (state: AppState) => void;
+}
+
+export interface State {
+  searchTerm: string;
+}
