+++ src/types/SearchResult.ts
@@ -1,0 +1,8 @@
+export interface SearchItem {
+  uid: string;
+  name: string;
+}
+
+export interface Props {
+  items: SearchItem[];
+}
